[TBL109] A pipeline to deploy cloudfoundry apps
For each repository there is a pipeline which starts when there is a
commit to a repository. An example pipeline looks like:

ras-party -> deploy to dev -> integration test -> deploy to test

If any step along that path fails the next task will not execute.

The integration test task is currently a placeholder for when the tests
are implemented.