Add known issues section (#12)
Creating a known issues section for adding issues that we have identified and the way to work around or resolve. Sometimes we will see an orange box on concourse and this means concourse has failed for some reason. The error we got was pq: current transaction is aborted, commands ignored until end of transaction block. This is a known issue and can be resolved by restarting the service